backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
idna==3.10
oauthlib==3.3.1
pillow==11.2.1
psycopg==3.2.9
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
ruff==0.8.0
social-auth-app-django==5.5.0
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
